+++ src/asphalt/mailer/mailers/smtp.py
@@ -2,64 +2,64 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Iterable
+from contextlib import AsyncExitStack
 from email.message import EmailMessage
 from ssl import SSLContext
 from typing import Any
 
-from aiosmtplib import SMTP, SMTPTimeoutError
-from asphalt.core import add_teardown_callback, get_resource_nowait
+from anyio import fail_after
+from asphalt.core import get_resource_nowait
+from smtpproto.auth import PlainAuthenticator, SMTPAuthenticator
+from smtpproto.client import AsyncSMTPClient
 
 from .._api import DeliveryError, Mailer
 
 logger = logging.getLogger(__name__)
 
 
-class SMTPMailer(Mailer):
+class BaseSMTPMailer(Mailer):
     """
-    A mailer that uses `aiosmtplib`_ to send mails.
+    Base class for mailers that use `smtpproto`_ to send mails.
 
-    The default port is chosen as follows:
-
-    * 587: if ``username`` and ``password`` have been defined and ``tls`` is ``True``
-    * 25: in all other cases
+    If ``port`` has not been provided, it is set automatically to 587 if
+    ``authenticator`` was provided, otherwise 25.
 
     :param host: host name of the SMTP server
     :param port: override the default port (see above)
-    :param tls: whether to initiate TLS using STARTTLS once connected (defaults to
-        ``True`` if ``username`` and ``password`` have been defined)
     :param tls_context: either an :class:`~ssl.SSLContext` instance or the resource name
         of one
-    :param username: username to authenticate as
-    :param password: password to authenticate with
+    :param authenticator: the authenticator to use for authenticating against the SMTP
+        server
     :param timeout: timeout (in seconds) for all network operations
+    :param max_concurrent_connections: maximum number of allowed concurrent connections
+        before calls start to block
     :param message_defaults: default values for omitted keyword arguments of
-        :meth:`~asphalt.mailer.api.Mailer.create_message`
+        :meth:`~asphalt.mailer.Mailer.create_message`
 
-    .. _aiosmtplib: https://github.com/cole/aiosmtplib
+    .. _smtpproto: https://github.com/agronholm/smtpproto
+
     """
 
-    _smtp: SMTP
+    client: AsyncSMTPClient
 
     def __init__(
         self,
         *,
-        host: str = "localhost",
-        port: int | None = None,
-        tls: bool | None = None,
-        tls_context: str | SSLContext | None = None,
-        username: str | None = None,
-        password: str | None = None,
-        timeout: float = 10,
+        host: str,
+        port: int | None,
+        tls_context: SSLContext | str | None = None,
+        authenticator: SMTPAuthenticator | None = None,
+        timeout: float = 30,
+        max_concurrent_connections: int = 50,
         message_defaults: dict[str, Any] | None = None,
     ):
-        super().__init__(message_defaults or {})
+        super().__init__(message_defaults)
         self.host = host
-        self.tls = tls if tls is not None else bool(username and password)
-        self.port = port or (587 if username and password and self.tls else 25)
+        self.port = port or (587 if authenticator else 25)
         self.tls_context = tls_context
-        self.username = username
-        self.password = password
+        self.authenticator = authenticator
         self.timeout = timeout
+        self.max_concurrent_connections = max_concurrent_connections
 
     async def start(self) -> None:
         if isinstance(self.tls_context, str):
@@ -69,39 +69,90 @@ class SMTPMailer(Mailer):
         else:
             tls_context = self.tls_context
 
-        self._smtp = SMTP(
-            hostname=self.host,
+        self.client = AsyncSMTPClient(
+            host=self.host,
             port=self.port,
-            tls_context=tls_context,
+            connect_timeout=self.timeout,
             timeout=self.timeout,
+            ssl_context=tls_context,
+            authenticator=self.authenticator,
+            max_concurrent_connections=self.max_concurrent_connections,
         )
-        add_teardown_callback(self._smtp.close)
 
     async def deliver(self, messages: EmailMessage | Iterable[EmailMessage]) -> None:
         if isinstance(messages, EmailMessage):
             messages = [messages]
 
-        try:
+        async with AsyncExitStack() as stack:
             try:
-                await self._smtp.connect()
-
-                # Authenticate if needed
-                if self.username is not None and self.password is not None:
-                    await self._smtp.login(self.username, self.password)
-            except Exception as e:
-                raise DeliveryError(str(e)) from e
+                with fail_after(self.timeout):
+                    session = await stack.enter_async_context(self.client.connect())
+            except OSError as e:
+                raise DeliveryError(
+                    f"Error connecting to {self.client.host} on port {self.client.port}"
+                ) from e
 
             for message in messages:
                 try:
-                    await self._smtp.send_message(message)
+                    with fail_after(self.timeout):
+                        await session.send_message(message)
+                except TimeoutError:
+                    raise
                 except Exception as e:
                     raise DeliveryError(str(e), message) from e
-        finally:
-            if self._smtp.is_connected:
-                try:
-                    await self._smtp.quit()
-                except (ConnectionError, SMTPTimeoutError):  # pragma: nocover
-                    self._smtp.close()
+
+
+class SMTPMailer(BaseSMTPMailer):
+    """
+    A mailer that uses `smtpproto`_ to send mails.
+
+    The default port is chosen as follows:
+
+    * 587: if ``username`` and ``password`` have been defined and ``tls`` is ``True``
+    * 25: in all other cases
+
+    :param host: host name of the SMTP server
+    :param port: override the default port (see above)
+    :param tls_context: either an :class:`~ssl.SSLContext` instance or the resource name
+        of one
+    :param username: username to authenticate as
+    :param password: password to authenticate with
+    :param timeout: timeout (in seconds) for all network operations
+    :param max_concurrent_connections: maximum number of allowed concurrent connections
+        before calls start to block
+    :param message_defaults: default values for omitted keyword arguments of
+        :meth:`~asphalt.mailer.Mailer.create_message`
+
+    .. _smtpproto: https://github.com/agronholm/smtpproto
+    """
+
+    def __init__(
+        self,
+        *,
+        host: str = "localhost",
+        port: int | None = None,
+        tls_context: str | SSLContext | None = None,
+        username: str | None = None,
+        password: str | None = None,
+        timeout: float = 30,
+        max_concurrent_connections: int = 50,
+        message_defaults: dict[str, Any] | None = None,
+    ):
+        port = port or (587 if username and password else 25)
+        if username is not None and password is not None:
+            authenticator = PlainAuthenticator(username=username, password=password)
+        else:
+            authenticator = None
+
+        super().__init__(
+            host=host,
+            port=port,
+            tls_context=tls_context,
+            authenticator=authenticator,
+            timeout=timeout,
+            max_concurrent_connections=max_concurrent_connections,
+            message_defaults=message_defaults,
+        )
 
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}(host={self.host!r}, port={self.port})"